feat: Add test for adding a song to an existing playlist
